Expanded jumbotron and moved register button to header
