+++ src/components/Hero/index.module.scss
@@ -1,0 +1,4 @@
+.jumbotron {
+  width: 40vw;
+  box-shadow: 0 0 30px 30px #000;
+}
